refactor:/components폴더 구조 변경
