+++ src/components/exhibition-gallery/threejs/Ceiling.tsx
