+++ src/components/exhibition-gallery/threejs/Floor.tsx
