+++ src/components/exhibition-gallery/threejs/ModalWrapper.tsx
