+++ src/components/exhibition-gallery/threejs/Painting.tsx
