+++ src/components/exhibition-gallery/threejs/Player.tsx
