+++ src/components/exhibition-gallery/threejs/Walls.tsx
@@ -1,5 +1,4 @@
-import React from 'react';
-import { WAllType } from '../../../types/gallery';
+import { WAllType } from '@/types/gallery';
 
 export default function Walls({ walls }: { walls: WAllType[] }) {
   return (
