+++ src/components/my-page/ExhibitionList.tsx
