+++ src/components/my-page/NewExhibitionBanner.tsx
